Feat/add exceptions

+++ sportslife-backend/src/main/java/com/grupo3/sportslife_app/service/UserService.java
@@ -9,10 +9,12 @@ import org.springframework.stereotype.Service;
 import com.grupo3.sportslife_app.dto.CreateUserDTO;
 import com.grupo3.sportslife_app.dto.NotificationDTO;
 import com.grupo3.sportslife_app.dto.UserDTO;
+import com.grupo3.sportslife_app.exception.UserNotFoundException;
 import com.grupo3.sportslife_app.model.GoalBoard;
 import com.grupo3.sportslife_app.model.SportRoutine;
 import com.grupo3.sportslife_app.model.User;
 import com.grupo3.sportslife_app.repository.UserRepository;
+import com.grupo3.sportslife_app.security.SecurityUtils;
 
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
@@ -32,9 +34,15 @@ public class UserService {
 
     @Autowired
     PasswordEncoder passwordEncoder;
+
+    @Autowired
+    SecurityUtils securityUtils;
     
     @Transactional
     public User create(CreateUserDTO userDTO) {
+        if(existsByEmail(userDTO.email())) {
+            throw new IllegalArgumentException("Email " + userDTO.email() + " already exists");
+        }
         GoalBoard board = new GoalBoard();
         SportRoutine sportRoutine = new SportRoutine();
         User user = new User(null, userDTO.name(), userDTO.email(), passwordEncoder.encode(userDTO.password()), board, sportRoutine, null, null);
@@ -59,10 +67,14 @@ public class UserService {
     }
 
     public User getById(Long id) {
-        return userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
+        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User not found"));
     }
 
     public void deleteById(Long id) {
+        Long userLoggedId = securityUtils.getCurrentUserId();
+        if (userLoggedId == null) {
+            throw new UserNotFoundException("No user logged in");
+        }
         User user = getById(id);
         userRepository.delete(user);
     }
